Allow user to specify an alternate iterationCount
Iteration count is hard-coded to 1024 in this file.  The change allows user to specify an alternate value to use (preferably higher).

Alternative implementations would be adding an additional constructor or changing parameters but I didn't know which method you'd prefer so I went with something that's not very intrusive by adding a method.