making unwrapSessionKey() protected instead of private
extending JceKeyAgreeEnvelopedRecipient (which extends JceKeyAgreeRecipient) and overriding extractSecretKey(), which already was protected to use some hardware-based key, I would like to reuse JceKeyAgreeRecipient's implementation of unwrapSessionKey(), but I can not, because it is private. Changing this to protected allow effective reusing of that implementation, instead of having to copy the functionality and even re-implementing it more complicated, as it uses other package-private methods, my code does not have access to.