Enable nohttp check during the build
HTTP is a plaintext protocol which means
that someone may be able to eavesdrop the data.
To prevent this, HTTPS should be used whenever possible.
However, maintaining using https:// in all URLs may be difficult.
The nohttp tool can help here. The tool scans all the files in a repository
and reports where http:// is used.

Updated http:// to https:// in URLs. Suppressed some findings.