Add factory method to instantiate a public subkey from a public key packet.
Unfortunately, the only public constructor for public keys does not set
subSigs to a non-null value. Since isMasterKey() derives its result
from the nullity of the subSigs list, it is not possible to create
public subkey objects without going a detour over the KeyRingGenerator.
This commit adds a factory method which sets the subSigs list to an
empty but non-null list.