Add method to join two copies of a public key
This method comes in handy, whenever a public key needs to be merged
with another copy of the same key, e.g. when consuming updates from
a key server.

This implementation does not do any verification on signatures and
blindly merges the data together.
Verification is left to the implementation, just like in other places
of BCs PGP implementation.