Expose PGPOnePassSignature.isEncapsulating()
This method exposes the nested flag of OnePassSignatures.
Since the naming was a bit confusing, I renamed the flag to
isEncapsulating.

Exposing this flag to consumers is required for consumers to verify
the signature properly. Before this change, consumers would need to
encode the signature and access the correct byte to check the flag.