Ant test output can now show incremental progress
Added ability to override properties in bc-build.properties
with a bc-build.user.properties file, which is ignored by git.
The intent is for bc-build.user.properties to hold user-specific
settings that are *not* committed to the repository.

A summary line for each test suite executed can be output by
adding this property to bc-build.user.properties:
  junit.printsummary: on

The default behavior remains unchanged, as the default value for
junit.printsummary is false.