Use Java's rotateLeft in salsaCore to increase likelyhood of inlining
We use BouncyCastle's scrypt implementation in
[Aegis](https://github.com/beemdevelopment/Aegis) (an Android app) to derive a
key from the user's password. We recently received [a
report](https://github.com/beemdevelopment/Aegis/issues/1024) from a user that
key derivation was taking very long on their device (on the order of 2 minutes).
The parameters we're using are: ``N=2^15``, ``r=8``, ``p=1``.

Personally, I'm only able to reproduce this when running the app with a debugger
attached on Android 13. In my case it takes 30 seconds. Since that user was
seeing this on a release build of the app, we figured this was worth
investigating. After some profiling, we noticed that a lot of time was being
spent calling the ``Integers.rotateLeft`` function in
``Salsa20Engine.salsaCore``. On a hunch I pulled a copy of that class and the
``SCrypt`` class into our codebase and replaced the calls of that wrapper
function by directly calling Java's ``Integer.rotateLeft``. With the patch
applied, it takes 3 seconds instead of 30 seconds. A 10x performance
improvement.

I suspect that previously the JVM was not inlining ``rotateLeft`` for some
reason, but with the patch applied, it does. With our scrypt parameters,
``Integers.rotateLeft`` is called 2^25 times, so it makes sense that this starts
to add up.

For now we're keeping a copy of the ``SCrypt`` and ``Salsa20Engine`` classes in
our codebase with this patch applied, but I figured I'd submit the change here
as well to your thoughts on whether this is something that should be upstreamed.