Add `len` parameter check and main loop changes in ConcatKDFGen
SP 800-56A Rev. 2 [1] is superseded by SP 800-56A Rev. 3 [2].
SP 800-56A Rev. 3 [2] doesn't include the **The Single-step**
**Key-Derivation Function** section anymore and the change-log states
that
> 34. Moved all key-derivation methods to SP 800-56C. Inserted a new
> section (Section 5.8.1) to describe how to call a key-derivation
> method and reorganized Section 5.8.

Following SP 800-56C Rev. 2 [3] this commit implements the changes and
will add conditional check for `len > 0`. It also brings the algorithm
more in line with how it is written in SP 800-56C.

[1]: https://csrc.nist.gov/publications/detail/sp/800-56a/rev-2/archive/2013-06-05
[2]: https://csrc.nist.gov/publications/detail/sp/800-56a/rev-3/final
[3]: https://csrc.nist.gov/publications/detail/sp/800-56c/rev-2/final

Signed-off-by: Burak Can <Fethbita@users.noreply.github.com>