Delay KangarooAbsorb until needed, to avoid 168 erroneous 0-byte-padding
This avoid too eagerly absorbing queued data, on queue fill-up, which would be wrong
if there were no more data to be written, and the 10*1 bit-padding should complete
within the same byte as the one which contains the start of the padding.
This bug would trigger when, e.g., a buffer of 166 bytes, and an empty personalisation,
was hashed, where the write of the "SINGLE" delimiter would fill the data queue, and
be immediately absorbed, followed by an explicit pad-and-squeeze, which would then
add another 168 0-bytes, with the final bit set. The correct would instead be to set
the final bit of the buffer after adding the "SINGLE" delimiter, and then absorb that.

Signed-off-by: Jon Marius Venstad <venstad@gmail.com>